Zero point, the beat of the heart and the rhythm of the world.
Add description.

This very first basic system, builds a quite minimal shell (zs), and an
utility to execute commands with superuser (su) rights (zsu).

The development route is to split the code, to very small libraries,
with a specific functionality.

Estimated time for the task completion and to a stable product, is
about eighteen months, working hard quite regularly.

  baba