z/zs: make the pipelines asynchronous.
Do not wait to collect the exit status from processes that are
part of a pipeline. Make something to ensure that no zombie processes
will left at the process table, during the execution of the pipeline.

This minimal shell, has the basic and the minimum functionality to be usefull,
minus the interactive part (but this is out of scope for now).