Add File command.
This is to explore the api. File have methods and are installed as:

  Type.method

in this case only a method for now:

  File.size filename

It can read the filename from stdin.

This looks quite consistent and maps to the library types, such
method names and argument. In this case the function signature is:

   size_t File.size (char *filename);