A very first draft of an application and abstraction layer.
We are using the Dir commands, as a vehicle to understand the needs.

The only worry is, if we are hiding too many details, because we do.
But these convenient abstactions, will trully speed up quite a lot
the development process. If we go to far, we can always revert them
in a state by parsing the units and substitute with the real code.

Hopefully the environment will be established and be appreciated, as
the clarity isn't missing.