Now the editor makes use of all our available libraries, so we can reduce significantly the compiled time.
This will allow also a separation of conserns, though we are also developing the conserns,
but this is not the point.

But because of this code simplification, we were able to  fix a couple of bugs.

We have to fix more though, or introduce new mechanisms to reach to the libved functionality,
but we are close.

But it was a nice thing, to be able to adopt almost 20000 lines of code, in a new system,
withing a couple of days of work, or a little bit more anyway, but not that much if you
count the complexity.