This is an initialization code to extend Sys Namespace, to get a bit more of control.
The details though are hiding within a commit with a lot of clean up code, so
we are missing some points here, as a commit could serve as a description and
a pointer, so it could be used more wisely.

Anyway. this also introduces the capability to set some environment variables, during
make invocation, which could be used by a compilation unit and its Makefile.