Editor and Visual (line|char)wise mode: Execute selected as a system command.
Editor: add :tty_screen command, to suspend the application temporarly and view
the original terminal screen. Usefull when is desired to view past messages on
the screen from previous executed commands output.

Replace the "free" method with "release", so to be consistent with the rest of
our world of libraries.