I: add malloced_string type.
These are allocated when literal strings are passed as arguments to C functions.
Untill now, the user was responsible to free them. Now they released automatically
at the end of the script, or at exit.

Note that passing literal strings to C functions is not supported upstream, since
this will introduce dependency to stdlib.