Added support for C-strings, by binding (almost all) the Cstring Type.
The same dangerous rules apply for this interface. That means that they
should be freed by the user, and that the user should know them well, e.g.,
their memory size.

Such a string can be allocated with alloc (size), or with cstring_new (size)
or with cstring_new_with ("string"). In the latter case the memory size is
the length of the "string" argument plus 1 byte for the nul byte.

Note that this, it won't never be a feature of tinyscript, which has no dependency,
and such feature will introduce huge dependencies. For us this is not an issue,
since we also have the interface for free, except malloc() and free().