This is to response to the last commit, with regards to the interpreter and the value representation.
It replaces intptr_t with a struct; with a type field to denote the type
of a union that holds the (three for now) the types.

These are a Double Type, a (char *), and a wide Integer Type, which is such
wide as ptrdiff_t, which it should be as ssize_t, and so handles size_t (if
i got them right), and it should be capable to hold a pointer  (the proper
type for pointer arithmetic actually).

These three types might be enough for us, as it also has this flexibility to
extend it in as many types you like and probably to all kind of fancy things
with it, by using it to set attributes. The reserved size of this union, it
is anyway and the size of the largest member type, which is double so it is
free.

This besides the obvious benefit, it brought also clarity in the code side,
since it is the easiest possible mechanism. At the same time though, added
a bit of overhead (possible more than a bit to be honest), because some BINOP
operations, like testing for equality of two objects, they need to handle the
cases. Before this was a plain known operation on two ints. Such operations
now (like a multiplication), they need to check both the types, to set the
proper type when it returns the result of the operation.

But this can  be turned into an advantage, because we could simple put some
logic to the switch() statement to the function of the operation (like an
addition), so it could handle strings for instance.

A branch using NanTagging, was abandoded due to techinally difficulties
faced (on 32bit), and to some low level bit technical details, and which
were important to fully understand, without paying a bit of pain when you
are trying to extend or modify the mechanism. So finally i realized that
this was a bit over my abilities and I gave it up with the idea. I think i
favor clarity over genious but complicated mechanisms. But yes it is really
magic, and we might find another chance to use it at some point, though it
is a bit of hardwarish, and complex for those who favor expressionism over
performance.

I mean it is the clarity of the intention that really matters, when writing
and reading code. Even the symbol names it is a factor, as they should have
clear relationship, consistency and speak about their underlying type, which
is most important for C, that is dealing with types that really matters to
get and deal with them right.

Though the API change was pretty invasive, the implementation finally was pretty
straightforward, because of this simplicity. The Nan Tagging method, it took me
two days of headache, with a couple of bugs, still left to handle.

This is passing all the semantic tests and valgrind reports a clean state.
Of course the C interface is now broken with this API, since the function
signatures they need to adjust. For this reason and because i do not want
to break the applications that uses the interpreter, i decided to add it as
a separate concept, as it actually starts to look a little bit more like
a programming language, and bind it to an own namespace named La (for Language).

Obviously there isn't any expectation other than Joy.