Initial support for C-string and Number arrays.
The declaration array syntax:

array integer varname[len] = 1, ...
array number  varname[len] = 1.1, ...
array cstring varname[len] = "string"

If the type is ommited, it defaults to an integer array.

This is incomplete, as it still needs to handle double free's cases.
It also needs to handle the length with a new syntax, though the current
it still works the same as it did.

Also tests should be adjusted to the new interface.