Lookup for operators in the global environment.
Allow anonymous functions to influence the outer scope througth their
arguments.

The previous commit, introduced this behavior for user defined functions,
which is an imitation of copy by reference, but only for memory types, like
strings or arrays.

That means that the value of integer or number type variable,  and which is
passed as function argument, can not be modified by the calling function.

However the real practical value for anonymous functions is questionable.