Add an ObjectType which is under development.
Use it and expose a fopen() and a fflsuh() wrapper to use FILE pointers in
the language. For now fclose() hasn't been exposed.

Use those file pointers to the print[ln] function. Expose also `stdout` and
`stderr` as the FILE *pointers of the standard streams. For now `stdin` is
not handled.

But the most important to keep here, is the way we handle (with a bit of hack)
the case, when function call results are arguments to functions. It was prettyy
difficult to find a way, so to make sure all the resources are being freed at
the right moment. Finally it was a couple of lines of code at the right place
though there isn't a build in stone certainity that this is sufficient. The
tests however are all passed without leaks and|or memory invalid accesses. It
took a bit of time, because attempts were ended up either without leaks,  but
with invalid memory access or the other way around. Just a reminder as how weak
is the underlying mechanism, though still goes on and on, and for some time now,
and which is probably sufficient even this state. Everything else is a gift.