Make the "+=" to work for strings also. If the second operand is an integer, then it appends the character if it is within the ascci range or else a multibyte sequence.
Switch the if/else[if] conditionals to switch statements in two of the important
in logic functions.

In IO.parse_escapes emit the double quote if it is escaped, but I'm not sure if it
is the right thing to do, or if it is the right place to do it unconditionally (the
mind is in a bit of confusion here).