Use the '+' operator for string concatenation.
Currently an integer is considered as a character, either as a byte if it is within
the ascii range or as a multibyte sequence.

There is some uncertainity, not because it could also be used to return this integer
as a string, but because this in C means another operation. Not sure though how far
we should follow with C.

Also, an internal detail, is the way to handle literal strings. From this commit we
just check if its symbol is NULL. In that case we release it.