Add loadfile(`fname`) function and `__loadpath` variable.
If `fname` is not an absolute path, then it is relative to the current evaluated
unit. If that fails, then it is relative to the current directory, else it is
relative to the `__loadpath` variable.

Add ECANNOTGETCWD error. The string error message that it produces is:
"Can not get current directory"