Initial draft of a more flexible test suite.
This migration revealed a couple of memory leak cases, that weren't handled
properly. Hopefully won't be much more, as the machine got too complicated
for a poor human mind to handle all those states. But it works! so far. So,
this is an important commit to check for back references.