Add anonymous arrays.
This allows simpler array declaration and assignment syntax:

  var ar = ["one", "two", "fortythree"]

alas with a price, as the function should do first a "light" parsing, just
to determinate the length of the array, and then the "real" evaluation.

But since we do not care much of speed this is fine, since it make much more
sense as it is an expected and stabilized syntax. Anyway it is enough speedy
for our tastes and tests!