Enhance ArrayType with a MapType.
This is a first draft, and even if the code might not handle all the cases,
we've done with the types.

This is good news, since we've reached this humble machine quite over of its
limits. But there are still many code variations and expressions, that are not
handled with the current code, and i'm not sure if we can, since the state is
quite quite fragile. Anyway this crazy adventure can be considered as completed,
as what lefts to be coded, are functions and operations on the types. But this
is another thing.