This is a first response to the last comment of the last commit.
As an already gain by trying to adjust some things, a map can handle now
properly user defined functions and C functions.

Also enhanced the typeof() function.

Some things might not work as it was intented, but the tests are ok and a
path module works flawlessly with some quick tests.