Add another form of the `for` loop, that works as an iterator.
For now it is implemented only for MapTypes.

Its syntax is common in other languages too, and is expressed as:

  for |k, v| in map { block }

In this case 'k' holds the key of the map, and 'v' its associated
value.

However, it might be needed to remove the posibility to set a field as
`private`, otherwise we might introduce complexity. For now there is
a flow in the logic.