Change the rather incosistent array declaration, to a more natural way that makes much more sense.
This feels that it should stay as it is as a final syntax:

  var integer[length] identifier
  var number[length] identifier
  var string[length] identifier
  var array[length] idsentifier
  var map[length] identifier

to create an array with a predefined length and with an appropriate type.

As a plus, parsing now it is a bit quicker, as it avoids many conditional
blocks and with a more direct call to the parsing function (at least one
less function call), but with an extra conditional block for a price.