Change the semantics of `lambda` functions.
Now require an argument list after their declaration, since they are called
immediately.

Release their resources after the evaluation. Because of that it is not longer
possible to assign them to a variable or to pass them as arguments to functions.
The func keyword should be used for those cases.