Move some tests into the semantics function, extend the std module, and also add the documentation for the next commit.
This is to avoid some noise in the next commit that has to be as easy to look,
if there is a need in time, and it is an exception to the rule.

Generally speaking, it is a bad habit to include unrelated changes in a single
commit and i do it all the time. But i do hate small commits. In fact, i want
the git history (that it is recorded in the log messages), to feel more like a
diary of activities, that also work as a documentation and as a recording of the
history of a project during its evolution.