This is a really really early draft of a chaining function mechanism, that the current function result, it is also the first argument to the next function in the chain.
It is accessed using a colon (':') instead of a dot ('.'), because of the different
semantics. It is also used in submaps, where it is pushed the submap as "this".

Here are some examples that shows the mechanism, and should owe to work as
intented.

  var sd = "97"
  var dd = sd:to_integer ()    => 97
  var ii = 111:to_string (16)  => 0x67
  var sa = ("10":to_integer () * 12 + 52 - 24) : to_string (2) => 10010100
  var vv = "asdf":eq ("asdf")  => true
  var vx = "asdf":eq_n ("asdfg", 5) => false
  var xx = "fa:fb:fc":tokenize (":") => ["fa", "fb", "fc"]
  var xa = [1, 2, 3]:len () => 3
  var xm = {"k" : 1, "l" : 2}
  var xk = xm:keys () => an unordered array of "k" and "l" as members.

This is to introduce the mechanism, to the and set to, development environment,
so it would be easy to work during its evolution. This it might take time to be
stabilized, not really the interface, as even now it seems to work, but mostly
to handle properly the memory resources. But really mostly to stabilize the underlying
semantics, that have to obey in without second thoughts expectations, but really
more about the way we'll to handle the inner chained environment, e.g., more
like a closure, so to avoid any side affects with the outer environment, and
which it looks that it might be the way to go.
So probably maybe will follow small commits. Just this in contrast to the conduct
that lied lazily in the very most previous commit.