Add the packman library and the corresponded Packman command, which are intented as an inner packaga managment mechanism.
This besides that at some point will be used to build the ecosystem,
it is perfect to stretch the underlying machine, because of its
complex (for a new language) expressions and mechanisms, like:

maps - submaps - evalfile - private/public scope - Type/New keywords,
importing modules - constants - setters/getters - break with an
explicit loop level - executing system commands, arrays ...

Tested with valgrind for memory leaks.

Add a specification to merely test the code, which is `lr' from
https://github.com/leahneukirchen/lr.

Haven't found the chance to try the git clone/pull code or|and the fetch
operations.