Allow lambda functions to be a part of a function sequence.
Allow multiply statements in a single line, without using a semicolon as
a separator. This doesn't always works, so if in doubt a semicolon should
be used. This change on semantics, will assist in a more natural code flow,
because sometimes it makes sence like:

 lambda (x) { if (x > 10) { return x / 2 } return x * 2} (10)

though it is still a bit hard to parse when you read it, but it is natural
when you write it. But multiply statements in a signle line, it is a bad
programming practice anyway.