Handle format() when is a member of a function sequence, since format isnot a proper function.
This now provide some inconsistency, since when it is called directly, any
function call (inside the string evaluation) it can accept literal double
quotes (though this doesn't look exactly right).

Now, all the memory resources are handled with the new function sequences code
(though quite possible there are unhandled cases, but at least are handled, where
this code is being used and with the way it is being used). All the old code works
as before the changes.