Add when/orelse conditional expression mechanism, which is executed inside of a function sequence call.
This is like an `if' with some differences.

A `when` expression, it awaits an input value.

A `when` expression, it returns a value.

And unlike `if` it supports only an optional `orelse` clause, as this is more like:

  (condition) ? expression : expression

in C language.

Inside the body of a `when` expression, a return statement it pushes a
value in the stack of the chain and returns control to the outter scope,
but it does not try to exit of a function.

A `break` or a `continue` statement is not allowed in a `when` expression,
unless is inside a loop of a `when` expression.

Generrally, `when` expressions and the function sequences mechanism, they
are trying a bit hard, to minimize any influences with outter environments,
though obviously this is not possible, as anything can be written inside a
body of a `when` expression or a function call. But at least this is the
intention of the mechanism.

When/orelse syntax:

  (expr): when |x| condition { body } [orelse { body }]

note that |x| is the symbol associated with the value of `expr', that can
be used in the condition or|and to the body.

Also note that `condition' is not obligated to be surrounded by parentheses.

Examples:

  12: when |x| x < 12 { return x + 22 } orelse { return x + 32 } => 44
  ("asdf" is "asdf"): when |v| v { return ok } orelse { return notok } => ok
  "asdf" : when |x| x:eq ("asdf") { return ok } orelse { return notok } => same
  var mp = {"k" : 1, "f" : func {return 1}}
  mp: when |m| m: len () is 2 { m.k = 10 return m } orelse { m.k = 20 return m} => new map
  var ar = [1, 2, 3, 4]
  ar: when |x| x:len () is 4 { x[0] = 2 x[1] = 3 return x } orelse { x[0] = 22 x[1] = 23 return x } => new array
  var s = null
  s: when |x| x is null { return notok } orelse { return ok }  => notok
    (note that, if in the above, we changed the condition and we ommited
     the `orelse` clause, like:
       s: when |x| x isnot null { return notok } => null

     as in this case, the value of `s' would be returned)

This is very experimental and commited for development reasons. The whole
`chaining` mechanism is experimental, but the trueth is that it got a bit
too attractive lately and it is a bit fascinated (as much you can get a real
value from a computer), when around chains make sure that you wont get far
in contrast.