Stabilizing function sequences and when expressions.
This is (almost) the end of features and it signals the stabilizing period.

This commit handles maps and submaps, when they are input for function sequences.

Tests and application code were checked for memory leaks or invalid memory
reads with valgrind. Some of them quite complicated, though still there have
to be unhandled cases, but we are close and we already went away too far, for
our mental capabilities and our expertise.