Add a file_copy() function.
This merely started to test the qualifiers mechanism, and ended as cp(1)
implementation, as a single library function.

Bind this function to the file module with the same name.

Use qualifiers to create the state of the file_copy() function, and pass
this state through designated initializers.

Much of the standard cp(1) options should be handled and they mapped with
the qualifiers as:

  all  -a|--archive
  force  -f|--force
  follow_lnk  -L|--dereference
  preserve    -p|--preserve
  recursive -R|--recursive
  update  -u|--update
  verbose  -v|--verbose

The semantics should match the same semantics of cp(1), though can not verify
all the cases.

This has been tested only in interactive sections. There is no documentation
nor tests yet.