On map operations and when the name of the key is not known and it is constructed at runtime, now a dollar ('$') expansion can be used.
  var m = {}
  var v = m.$("new")
  var keyname = "new"
  m.$(keyname) += 1

Previously, a standard function had to be used and in the last statement two
C function calls, one to get the variable value and one to set the value.
Plus it was too verbose and a little weird for the established syntax.

This is the first time that this mechanism is being used, and actually is
more than a simple expansion, as it is a full expression. Not quite sure
about the syntax, though the parentheses they avoid ambiguity, while without
them the inner code would be complicated, and without the same clarity at the
end.