Add `break if[not]` and `continue if[not]` single boolean statements, that are executed when the evaluation of the expression that follows is true or false respectively.
Syntax:

   continue if (expression)
   continue ifnot (expression)
   break [count] if (expression)
   break [count] ifnot (expression)

In absence of a known term, as this is unique (from what I know), we can
describe this concept as a "Boolean single statement" or something like this.

It accepts only a single expression to evaluate, and it doesn't accepts clauses.
It works both for `if` and `ifnot`.

Finally, it has to be seen the same way as an `else if`, so an `if` or an
`ifnot`, should follow `break` or `continue` to the same line.

(notes: that this should execute faster, than with an equivalent traditional
   way and order. That is because the inner implementation, as an `if[not]`
   conditional, creates a function state everytime it is called, as it has
   to manage symbols scope, so it has to allocate and deallocate memory,
   plus it has to parse the code with a way that can interpret correctly
   `else if` and|or `else` clauses, and this is costly. In contrast only
   a single evaluation taking place, as there is no expectation for other
   options to look forward.

   Also note that it is believed that this is a more natural mind flow,
   as you might feel it (albeit with some bits of attention), when you
   will try to spell it, either loudly or in your mind.
   Here I'm feeling that I'm pulling something (air), or is like pushing
   back a thought, that seems that it is holding back than to continue
   with the flow, even for a negligible fraction of time) when I say:

     "if condition then break" in contrast with "break if condition"
)