Add an if as expression mechanism.
With this an if or an ifnot, can be used to evaluate any expression.

This is a very first draft, so quite probably there are unhandled cases,
though they have been tested as function arguments, in variable assignment,
as operands to binary operations, to get the index of an array, to access
a Map key with a dollar expressions (that means quite much all the set of
expressions).

However, because we want to support a very humanish way to write, and this
has a price to complication, then possible there might exist consuming bugs
mostly for un-evaluated clauses next to true. That is because we really want
to avoid ending such expressions with another way other than the established
(a semicolon, a new line, or the context). And in such expressions it is easy
to miss the end of an expression, or it is quite easy to end prematurely.
Otherwise a clear defined end, had to be used to handle them, like:

  if cond then do_something orelse do_something else end

  or like sh does:

  if cond then do_something orelse do_something else fi

  or other keywords like endif, e.t.c.

A bit of a little bit pain than it should, just to get it right, but with
a big reward at the end, just to give an end on ends.

It uses the exact syntax and semantics for if/then/orelse single statements,
only that it evaluates expressions instead of statements.