This has been started just to propagate errors from functions that can return fatal errors, even for those that is unlikely. This has a price but that is the proper thing to do.
But ended up to replace code routines with macros, just to give a (obvious)
boost to the clarity of the logic and to code intentions, though some C purists
may disagree a bit because details are hiding. But those are pretty much the
exact same patterns/tasks/requests all over the world. BTW, i hate caps in
the code and sparingly use macros, for this same reason, but the trueth is
that is also and the nature of this codebase, that may even require such a
bad programming style. The trueth is also that now they is  much easier to
follow the logic and the intentions.

Anyway, the diff for this commit is 7000 lines long but few.