This is to replace "when" expressions with "if" expressions.
And it does, but it lefts too much "when" code inside, but it had to
be committed because it went tututu far, and it is easy to loose control,
if you miss a point that you haven't thought about it.

Especially if you think, that it is really the first time that it has
been confirmed that does the right thing. But not well tested, as might
(and quite possible) there are unhandled cases.

The other cause of this, might is the most important point untill now for
this project. It defines a bare minimum set of semantics and syntax, that
can be used by any with a will other interpreted languages. Something like
a zero point, that can be supported by most if not all. The rest will be
a plus ('+'), and again with well defined and established concepts by now.