Development of backquoted strings.
Such strings can have attributes, that tune the behavior, such as 'S' that can
strip the leading whitespace. This take an option, which is the max number of
whitespaces, and follows 'S'.

Hasn't ended up to semantics yet, because there has to be a way to enclose those
strings in a pair, so they can be nested, without the need to double or escape
multiply times the backquote.