Add File.symlink command.
Abstract the details for getting qualifiers options from modules.

As a generic comment, make module functions to behave as their correspnded
commands. The idea is to use the operands as their basic arguments, and use
qualifiers to extend and tune their vehavior accordingly, without the need
to change their signature, so to gain future extensebility from one hand,
and the desired standardization from the other.