Add Crypt command, with three symbolic links that point to md5sum, sha256sum and sha512sum.
The code in this unit, revealed a bug in print[ln] function, that fails to
consume a closed brace '}' in interpolation syntax or to match a closed paren
')' in calls with a single object, and when the expression evaluates to function
reference and the function prints the pointer address.
Consuming or unget the extra token, breaks format() expressions, inside other
format() expressions, and calls to print[ln] functions.
This is as a reference to look, at the time it will be attempted to fix the
drain main brain again.