Move the random code to a library as it may be used by others.
Handle chained functions when getting the length of an array at
the initialization.

Handle conditional expressions when they don't surrounded by parentheses.

Handle user defined functions when they are called in a chained function sequence.

Handle "std[out|err|in]" from releasing when are assigned to a variable.

When getting a C function address without calling it, consume an
extra token. This might fix the print[ln] functions with the issue
mentioned at 531d47698f10f59ede7db0865dceaf2dda961975. This as a place
to look.

Look for '}' at the end of an interpolated expression.

Handle more properly [C|U]FunctionTypes. Still a mess at points though,
that is present almost from the very begining, when there wasn't an interface
to follow and stable routines.

Realize the fact that qualifiers, it is just one pending set of a map,
that are reinitialized at every function call. Thus the functions
that they use the qualifiers interface, should use it and get them early as
they might change with a new function call. I'm not sure this is
quite right.

Last, comment out BINOP code that seems it is|was never being used, and which
it might be the one piece that i never understood the intention.
So this is a pointer for the future, if something goes wrong and
proved wrong.