This is the development code for the postfix ++,-- operators. And this because there is a story about this, that envolves also C but at the end it's not the point.
I thought first and there is a comment in the Language document, that the postfix
operations were easier to handle. That was wrong impression at least for the
map and array member cases, because the mechanism of those types doesn't about
the associated symbols, so the information could be taken reliable only by using
pop/push operations on a structure specific for this, rarely used ++,-- postfix
operation. So it ended up as a hack, by doubling just the call.

But again its'not the point. The point is that beginning from the previous commit
the function chaining can modify the outter environment, by using those ++,--
operations on an associated symbol. At least as now, there wasn't an obvious way
that a function call in a sequence could influence the environment.

And it's not that there is an ambition that will ever even attempt to apply
functional concepts in the future, as there isn't the singlest knowledge about
functional programming. I guess operations on an anonymous value with a series
of expressions that cannot influence at least the outter environment, it maybe
qualified as functional.

The thing is that such operations are very handy in a loop and when you want to
advance the index of an array, without the need for one more statement.