This is to handle also prefix operations for statements, though they don't really have a usage as they do what postfix operations do, but just for the sake of completeness.
With regards to the previous commit. Postfix or prefix ++,-- operations
it's not the only obvious way to modify the outter environment in chained
functions. Any function in the sequence can also influence the environment,
and there is no way to prevent them from doing it so. Ideally only functions
that could be declared as pure should be members of such sequences. In my
mind that means functions, that in their calculations they don't depend on
an external state.

Right now, the first value from the environment in sequences of expressions,
is copied so it can not be modified during execution. Also the last value
that is produced returns to the environment as an anonymous value, that means
without any association with a symbol. So we are close. Though they do really
make sence only if it was an intelligent machine behind, that could benefit
from a functional environment. From the few this mind can handle in someway
about functional environments, is that the idea is that nothing can go ever
wrong, if the first time of the execution with some input succeeded. It will
always do, since it doesn't depend in nothing but internal state, and so it
will always return the same result.

Anyway, it is doubtfull the future of those prefix/postfix operations, as it
much code for little benefit, of which some may have side affects. Plus the
prefix handling is quite bit of a hack and fragile, as the machine isn't so
capable for a more deterministic way. For instance this is misinterpreted:
  ++v ++v
Okey for a reason, but it should be possible, though two statements in a single
line, without a semicolon at the end, it is not guarranteed, but they work almost
always without a semicolon.

But they are handy. And we can always disallow them where we do not want them.
Plus we are close to our brain limits and with almost no ambition. Lets the time
to decide.