Handle recursive functions by using the self keyword to call the own function.
In this case a tail call optimization it is performed. If the statement it is
not a tail call, then a syntax error is raised:

  func rec (n) {
    return n ifnot (n)
    return self (--n)
  }

  println (rec (2000000))

Now both the logical OR and AND are short circuit operations. That means, that
if an AND operation is false, evaluation stops and the condition is false.
If an OR operation is true, evaluation stops and the condition is true.