Allow loops with single statements that fit in a single line:
In this case the braces are ommited:

  while (true) break
  or
  while (true)
    break

But no more than a single statement that can fit in a single line, and which
may have a loop that produce another loop, as long all those they fit in
a single line. That means few, usually one.

Likewise do that for functions.
Currently the body can contain multiply statements, and evaluated correctly
as long they have clear semantics and which they usually have. But there are
cases that multiply statements in a single line without explicit semicolon at
the end, they produce unintented evaluation. For that reason only a single
statement can be guarranteed.
For loops a warning (for now) is raised if it follows a token that doesn't denote
the end.

Do not walk too far behind the current token, when printing an error message.
Many loops and conditionals, they make up indepented strings for evaluations,
that do not have relation with the original Parsing Pointer Stream.

Avoid an erroneous warning on break statements. Do that by if it quarks
like a duck and walks like a duck then it is a duck way.

Fix by returning the error value instead of OK, when interpreting break
statements into if/then/orelse.
Also return when EOF is encountered at the end of an if/then/orelse.

Finally handle function signatures that continue to the next line. Do it
only when a comma is followed by a new line token.