Remove the restriction that map Keys should be only strings.
But keep the existing capability to maintain compatibility with json.

Also strings can contain any byte, even whitespace.

An identifier though can contain only the characters from the set [_a-zA-Z0-9].

Also a dollar expression can be used to declare a key with an expression.

The same three methods can be used also to get/set a map field. Obviously
a key that has some characters that are not in the map key identifier set,
then it should be accessed either as a string or with a $(expression).

The length of a key raised to be 255 bytes.