Add File.remove() in libfile, File.remove command and File.remove() in file-module.
The code, implements force, verbose, interactive and recursive options.
It nevers follows a link.

That reminds me, the spring of 2010 while learning the S-Lang programming
language (my first code in a proper language in fact), the implementation
most of the coreutils.

So it seems we didn't walked too far pilgrim!!! as we still do the
same thing again.

Probably the one difference that might makes a difference and
might have a point - and it has in any case - it is that we do it by
using our own tools (the underlying C code (much of it implements
even low level based on stone libc functions!!!) and even our own interpreter).

But the trueth is that it looks that we come from backwards, pretty much
as a circulation. The trueth is also the we walked the path. Time will tell the trueth?