Add File.rename() in libfile, File.rename() in file-module and File.rename command, plus a File.move command.
The code, implements force, verbose, interactive and backup options.
The backup option it works only for regular files.

That reminds me, the spring of 2010 and my first patch against the mv script
of the S-Lang distribution.
So it seems we came back from where we came from. Time is an illusion, or we
are turning around in circles repeatedly. Or we have a plenty of time. Or we
just walked too far, in an uncaptured amount of sence of time. Or just there
is no time. But since its the time that gives shape then there is no space.
In any case, its about time.
Or it is just an impression anyway. Or it is the impression for time.