Make the shell and the interactive language session, a little bit more practical and useful to use, by introducing completion mechanisms.
Still are primitive interfaces though, as they have to be for their purpose. The visual
interface offers a much more rich and flexible completions.

Bump the revision at 0.3, as since now it is possible to run the system under an isolated
environment, which fulfills the second aim.

Probably though, and as we are in the middle of the journey, it is time to go back to
the roots for a while.