Add Bcrypt.
But before I go, this enhances this system shell, with tab completion and word
expansion.

To be able to support tab completion at any point of the cursor in the line, we
had to break linenoise API - structures, function signatures and exposed interface -
because upstream triggers completion only when the cursor is at the end of the line.

Also we've added two more callback functions. The first is called at the received input
and before any processing, and the second on a carriage return. The first permitted us
to do an almost overzealous and a bit agressive interaction, and when there is a chancce.
And this is a tendency, as this it might be the exact point of an interactive session
with an interpreter.

For now expansion happens through wordexp(). This is very convenient but also this comes
with some caveats, mostly because the implementations across C libraries differ. Most
of them just fork and let "/bin/sh" to do all the expansions, but not glibc and which
its implementation is almost the size of this shell itself. So at some point we should
do the expansion ourselves. In any case probably we might need a "/bin/sh", though
we have the freedom to ignore it, given all the (realized by all) quirks and the fragility
of the language.